% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nma.R
\name{extract_nma_estimates}
\alias{extract_nma_estimates}
\title{Extract parameter estimates as a dataframe}
\usage{
extract_nma_estimates(nma)
}
\arguments{
\item{nma}{multinma model}
}
\value{

}
\description{
Extract parameter estimates as a dataframe
}
