% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pw_studies.R
\name{find_pw_studies}
\alias{find_pw_studies}
\title{Find studies for pairwise comparison}
\usage{
find_pw_studies(g1, g2, dat)
}
\arguments{
\item{g1}{First comparator for meta-analysis: int}

\item{g2}{Second comparator for meta-analysis: placebo}

\item{dat}{Cleaned nma dat with study and intervention renamed}
}
\value{
Vector of study ids.
}
\description{
Filters dataset from input data (with study and intervention relabelled) to
studies that have arms for each of the two given comparators.
}
