% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_multinma.R
\name{forest_multinma}
\alias{forest_multinma}
\title{Multinma forest plot}
\usage{
forest_multinma(
  mod,
  model_type,
  dir,
  font_size = 15,
  this_class = "no class filter",
  or = TRUE
)
}
\arguments{
\item{mod}{A \link{multinma} model.}
}
\value{

}
\description{
Multinma forest plot
}
