% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_nma.R
\name{get_nma_dat}
\alias{get_nma_dat}
\title{Get NMA dataset}
\usage{
get_nma_dat(dat, ...)
}
\arguments{
\item{dat}{Currently mod_dat target}

\item{...}{Filters}
}
\value{

}
\description{
Get NMA dataset
}
\examples{

}
