% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pw_studies.R
\name{get_pw_dat}
\alias{get_pw_dat}
\title{Get pw data}
\usage{
get_pw_dat(dat, outcome, timepoint, g1, g2)
}
\arguments{
\item{dat}{Input model data \link{mod_dat}.}

\item{outcome}{Outcome pain_sub, pain_mod, etc.}

\item{timepoint}{Type of timepoint: post_int, change_score, etc.}

\item{g1}{Intervention of comparison.}

\item{g2}{Placebo or other ref comparator.}

\item{type}{Type of comparator: ad, etc.}
}
\value{
Wide dataset with placebo labelled as control.
}
\description{
Using \link{find_pw_studies}, identify the number of studies.
}
