% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_g.R
\name{hedges_g}
\alias{hedges_g}
\title{Hedge's g}
\usage{
hedges_g(mean_1, mean_2, sd_1, sd_2, n_1, n_2)
}
\arguments{
\item{mean_1}{Mean first group}

\item{mean_2}{Mean second group}

\item{sd_1}{standard deviation}

\item{sd_2}{standard deviation}

\item{n_1}{sample size}

\item{n_2}{sample size}
}
\value{

}
\description{
Hedge's g
}
