% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_forest.R
\name{hpp_forest}
\alias{hpp_forest}
\title{Create a forest plot using mod key inputs}
\usage{
hpp_forest(mod, mod_type, dir, font_size = 15, this_class = "no class filter")
}
\arguments{
\item{mod}{A \link{multinma} model}

\item{mod_type}{lor or smd}

\item{dir}{direction}
}
\description{
Create a forest plot using mod key inputs
}
