% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_forest.R
\name{hpp_forest_labs}
\alias{hpp_forest_labs}
\title{Label a forest plot}
\usage{
hpp_forest_labs(outcome, timepoint, mod)
}
\arguments{
\item{outcome}{outcome}

\item{timepoint}{post_int etc.}

\item{mod}{nma model}

\item{type}{ad etc.}
}
\value{

}
\description{
Label a forest plot
}
