% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp_themes.R
\name{hpp_tab}
\alias{hpp_tab}
\title{Theme across all tables}
\usage{
hpp_tab(gt_tab, vertical_divider = NULL)
}
\arguments{
\item{vertical_divider}{Dashed vertical line indicating row identifying columns}
}
\description{
Theme across all tables
}
