% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dat.R
\docType{data}
\name{mod_dat}
\alias{mod_dat}
\title{Model data created with targets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3135 rows and 75 columns.
}
\usage{
mod_dat
}
\description{
Model data created with targets
}
\keyword{datasets}
