% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnt.R
\name{nnt}
\alias{nnt}
\title{Number needed to treat}
\usage{
nnt(or, acr)
}
\arguments{
\item{rd}{risk diff calculated with \link{calc_rd_par}}
}
\value{

}
\description{
NNTs are always rounded up to the nearest whole number.
}
