% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnt_results.R
\docType{data}
\name{nnt_int}
\alias{nnt_int}
\title{NNT data for outcome, timepoint, and intervention}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 186 rows and 4 columns.
}
\usage{
nnt_int
}
\description{
NNT data for outcome, timepoint, and intervention
}
\keyword{datasets}
