% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_key.R
\docType{data}
\name{outcome_key}
\alias{outcome_key}
\title{Outcome information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 6 columns.
}
\usage{
outcome_key
}
\description{
Outcome information
}
\keyword{datasets}
