% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_forest.R
\name{pw_forest_labs}
\alias{pw_forest_labs}
\title{Label a forest plot}
\usage{
pw_forest_labs(outcome, timepoint, intervention, dat, mod)
}
\arguments{
\item{dat}{that goes into \link{pw_forest_plot}}
}
\value{

}
\description{
Label a forest plot
}
