% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_forest.R
\name{pw_forest_plot}
\alias{pw_forest_plot}
\title{Meta-analysis forest plot}
\usage{
pw_forest_plot(mod, dat, m_type, dir, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
First group is treatment group, second control.
}
