% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_int.R
\name{pw_int}
\alias{pw_int}
\title{Get set of pw comparisons for a dataset}
\usage{
pw_int(this_dat, this_outcome, this_timepoint)
}
\arguments{
\item{this_dat}{pw data}

\item{this_outcome}{outcome}

\item{this_timepoint}{timepoint}
}
\value{

}
\description{
Get set of pw comparisons for a dataset
}
