% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_nma_forest.R
\name{pw_nma_forest}
\alias{pw_nma_forest}
\title{Comparison forest plot of ma and nma}
\usage{
pw_nma_forest(dat)
}
\arguments{
\item{dat}{Nma and pw data}
}
\value{

}
\description{
Comparison forest plot of ma and nma
}
