% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob.R
\docType{data}
\name{rob}
\alias{rob}
\title{Risk of bias}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 805 rows and 5 columns.
}
\usage{
rob
}
\description{
Risk of bias
}
\keyword{datasets}
