% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.R
\name{sof}
\alias{sof}
\title{SOF}
\usage{
sof(mod, outcome = "antidepressants", pop = "adults with chronic pain")
}
\arguments{
\item{pop}{}
}
\value{

}
\description{
Based on \href{https://www.sciencedirect.com/science/article/pii/S0895435618303172?casa_token=VOokqa1sWuIAAAAA:pEXsLysLZybSApmf9u9pJ21Ze3J_HZLwktAAjZ-doSZy2l5lG_Fca3PuV-2ul1mM46INPsAhbA}{this ref}. There are three sections: upper, middle and lower.
}
\details{
Upper: PICO components + network plot
Middle: 8 columns
Lower: NMA terminology and abbreviations
}
