% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof_acr.R
\docType{data}
\name{sof_acr}
\alias{sof_acr}
\title{ACR calcs for outcomes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
sof_acr
}
\description{
ACR calcs for outcomes
}
\keyword{datasets}
