% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.R
\name{sof_meta_tab}
\alias{sof_meta_tab}
\title{PICO Summary of outcome component}
\usage{
sof_meta_tab(
  nma_mod,
  interventions,
  population = "adults with chronic pain",
  comparator = "placebo"
)
}
\arguments{
\item{mod}{}
}
\value{

}
\description{
PICO Summary of outcome component
}
