% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.R
\name{sof_results_dat}
\alias{sof_results_dat}
\title{Intervention results}
\usage{
sof_results_dat(nma_mod)
}
\arguments{
\item{nma}{nma mod results metatab}
}
\value{

}
\description{
Intervention results
}
\concept{sof}
