% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sof.R
\name{sof_results_tab}
\alias{sof_results_tab}
\title{Table of intervention estimates and ranks}
\usage{
sof_results_tab(results_dat)
}
\arguments{
\item{results_dat}{from \link{sof_results_dat}}
}
\value{

}
\description{
Table of intervention estimates and ranks
}
