% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_key.R
\docType{data}
\name{timepoint_key}
\alias{timepoint_key}
\title{Timepoint key from targets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
timepoint_key
}
\description{
Timepoint key from targets
}
\keyword{datasets}
