% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_key.R
\docType{data}
\name{type_key}
\alias{type_key}
\title{Type key from targets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
type_key
}
\description{
Type key from targets
}
\keyword{datasets}
