% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable_observations.R
\name{viable_observations}
\alias{viable_observations}
\title{Function to extract viable arms}
\usage{
viable_observations(hpp_df)
}
\arguments{
\item{hpp_df}{A subset of \code{w_obs}.}
}
\description{
Takes set of observations and returns only those with at least two observations
per study.
}
