% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_forgroup.R
\name{rep_forgroup}
\alias{rep_forgroup}
\title{Title}
\usage{
rep_forgroup(
  rep_outcome,
  rep_timepoint,
  rep_intervention,
  rep_comparator,
  rep_subgroup,
  rep_rma_prefix,
  rep_pw_mod,
  rep_pw_escalc
)
}
\arguments{
\item{rep_pw_escalc}{}
}
\value{

}
\description{
Title
}
